initial work on a built-in user manual
After yet another round of "Do we have documentation? Where?", create
some machinery for generating a user manual from content that actually
lives in our source tree. This way we have a better chance of keeping it
up to date and in sync with the code (even in case of major feature
changes, since the doc changes can accompany the code changes). The
custom generation script allows us to hook into reST's parsing in order
to define our own additions (e.g. we add a convenient way to reference
application icons that also automatically copies used icons to the
documentation output directory).

This includes an initial, woefully incomplete version of a vsPlay user's
manual. (Having a "complete" manual is likely to be a long and ongoing
process, and doing it in one commit is only reasonable by never
committing any intermediary versions. This way we at least have the
machinery and a demonstration of the same committed.)

Note that enabling this requires Python and python-docutils.