Simplify vgImage -> vtkImageData conversion
Avoid use of VTK operations and always do direct memory copies to
convert from vgImage to vtkImageData. It appears that the "simple"
conversion using VTK operations is in fact not so simple and is doing
something that makes the conversion horrifically slow compared to the
brute force byte level copying that we fall back on when we have strange
data alignment. Therefore, rip out the plane-interleaved path entirely
(falling back to the byte-by-byte copy instead) and replace the
"trivial" case with copying by scanlines (which is the best we can do
given the need to Y-flip the data, and is also optimal in the case of
padding at the end of scanlines which previously fell back on the worst
case path).

Also, fix the worst case path to work with any number of channels
(planes) rather than assuming 3. This allows us to entirely remove the
number of planes check (aside from the is-positive sanity check).