rename 'zoom on target' action
Rename the 'zoom on target' action and related implementation details to
'focus on target'. The 'focus' verb is preferred as more inclusive, as
the focus may include panning as well as zooming, and may be ongoing.
Add the action to the video menu to avoid it being available only via
the tool bar. Improve the tool tip text. Add documentation for the
'focus on target' action to the user manual.

(Note: this is based on a commit that was originally renaming the older
'follow event' action, which was independently renamed to 'zoom on
target' as part of a different set of changes that were included in the
most recent public release updates.)