Use VTK 6.2
Modify dependencies to require exactly VTK 6.2. Add link to new VTK
component vtkRenderingContextOpenGL, which fixes a "no override found
for 'vtkContextDevice2D'" error that causes us to crash at startup.

VTK has made breaking changes between each of 6.1, 6.2 and 6.3. Picking
one and enforcing it both reduces user confusion (avoids issues caused
by a "wrong" VTK version by making it explicit which one we expect) and
avoids trying to work around the chicken-and-egg problem of needing to
ask for different components depending on the VTK version when we need
to first find VTK to know what version was found in order to make the
build adaptable to the version of VTK in use.

For a description of the VTK 6.1 -> 6.2 breaking change, see also:
http://public.kitware.com/pipermail/vtk-developers/2014-July/030648.html