Create KWIVER In-Process query backend
Create a new query backend that speaks to a KWIVER pipeline. This is the
first step (at least from the ViViA side) in creating a full end-to-end
system using all Kitware software. For now, this should do a passable
job of interfacing with the KWIVER systems.

For now, the backend is optional. Eventually, we may want to start
migrating ViViA over to using KWIVER data structures internally, rather
than the historic ones we use now. (This, however, will involve
significant refactoring and will represent a massive break in
compatibility with old data and external backends.)