Port track_oracle plugin to KWIVER
Port the vsPlay track_oracle plugin to support use of the KWIVER version
of track_oracle. This makes it possible to use the plugin with the
publicly available version of track_oracle (which is in KWIVER).

In theory, these changes are backwards-compatible, i.e. will still work
with the old vidk track_oracle, although this has not been tested, and
we may eventually remove that support, since KWIVER is the path of
active development.