Update use of GeographicLib
Fix incorrect case of GeographicLib's name in the call to find_package;
the old case was partly working because GeographicLib uses the
'<lower>-config.cmake' naming convention, but proper case for the
directory name. As a result, CMake would only be able to find the
package when given the full path. Using the correct case allows CMake to
find GeographicLib when it is installed to a path that is searched by
default (e.g. via distro packages).

Also fix an outdated use of GeographicLib's API, and switch to using
imported targets to consume GeographicLib.