Add TOC support to vtkVgTrack
Add support for dynamic track object classification to vtkVgTrack. For
now, this is in addition to the hard-coded P/V/O classification that
already existed, although the latter is now deprecated and ideally will
be removed eventually.

Note that this uses integers as type identifiers. It is intended that
the mapping of these to human-readable types is handled at the
application level.