Implement QSettings-based vpProject parser
Implement an alternate parser for WAMI Viewer projects that is based on
QSettings (and thus has no additional dependencies) rather than VidTK's
config_block. This makes it possible to parse projects when VidTK is not
available.

Note that the projects are similar, but not 100% compatible. In
particular, AOI's are stored using a settings group, multi-value
fields require commas between the values, and the track override color
uses integers (0-255) instead of reals (0.0 - 1.0). The latter is due to
use of vgColor for reading, which makes it consistent with how colors
are read and written to application settings.

Also, reading the image-to-GCS matrix is not yet implemented.