Implement non-VidTK track reader for WAMI Viewer
Create an I/O implementation for WAMI Viewer that uses exclusively the
data frameworks (vdfSourceService) and does not depend on VidTK in any
fashion. (vpVidtkFileReader also uses vdfSourceService, but had a
fallback to VidTK's track reader process, and the retrieved tracks were
first wrapped in VidTK tracks; this new implementation cuts out that
step.) For now, only initial track reading is supported. Events,
activities, and importing are not supported.

This implementation also does not support the various ancillary files
that have sprouted up, but that is probably The Right Thing To Do
anyway; if we continue to support such things, they should be handled at
the reader plugin level, rather than partly at the application level, as
is being done currently.