Implement coloring by TOC, using type registry
Modify vtkVgTrackRepresentationBase to support coloring by a track's TOC
(rather than PVO), by using a vtkVgTrackTypeRegistry to supply the color
information. Modify WAMI Viewer to use this mode.

This deprecates most of the old PVO-based stuff in the representation,
but does not yet remove it, as doing so would require non-trivial
changes to some of the query video classes.