Refactor image conversion
Create a new helper class to fill a vtkImageData from an arbitrary
buffer. Refactor existing conversions from vgImage and in the
vtkVgSGIReader to use this new type. This replaces the pipeline based
mechanism that was being used in vtkVgSGIReader (which has serious
performance issues; see also the preceding commit) with a must faster
code path that performs direct copying from the input buffer to the
output vtkImageData.

This really should be done for vtkVgImageSource also, but that code is
somewhat esoteric and lacking in readily available data to exercise it,
so I'm leaving it alone for now.

Y-flip is accomplished with the new vtkVgImageBuffer by the clever
expedient of lying about the Y-stride. This shifts complexity from
computing the pixel- or scanline-offsets to computing the address of the
first pixel (so, about a wash) while allowing the caller to decide if it
needs Y-flip or not (which is important because SGI is already using
Y-up, and should not be flipped).