Tweak how region map is constructed
Modify vpFileTrackIOImpl::ReadRegionsFile to insert regions directly
into the correct region set in the overall map, rather than keeping an
intermediate set of regions for the current track. This is theoretically
less efficient (though the difference is likely negligible in practice),
but the resulting code is simpler, and it is robust against interleaved
tracks.