Refactor container accesses
Add some helpers in vpVidtkTrackIO.cxx to get values from an associative
container (these are moral equivalents to qtGet, but with some extra
features), that allow us to drastically simplify a recurring pattern
where we need to look up an element in a container, where the element
(and sometimes the container) may not exist. Refactor a bunch of places
that were just crying out to use range-based for.

Overall, this makes the affected code much more readable.