Use fletch
Fletch is part of the KWIVER (http://www.kwiver.org/) ecosystem, of
which ViViA is also a part. The purpose of fletch is to provide external
third-party dependencies required by KWIVER software. Previously, we
have had an implicit dependency on fletch, in that some or all of
ViViA's third-party dependencies would typically be provided by fletch,
with the developer manually specifying their paths. Adding fletch as an
optional dependency allows us to discover those dependencies which are
provided by fletch without additional user intervention, which makes it
easier to configure ViViA.