Rewrite vpFileDataSource using Qt
Rewrite the internal logic of vpFileDataSource using Qt, rather than a
mix of STL, VTK and kwsys, and refactor the API to use Qt conventions.
Update (most) users accordingly. This is incomplete, as vpFrameMap has
not yet been adapted, in anticipation of that class receiving a similar
overhaul.

This will make some coming work easier, and the implementation is
cleaner. (In particular, a totally useless mutex has been removed.)