Refactor vpFrameMap
Overhaul vpFrameMap, making use of Qt data types in much the manner as
the previous commit. This brings some increased consistency in the use
of signed rather than unsigned int as a frame identifier. Also, clean up
and fix use of synchronization in the background thread (which was
broken before!).