Implement (WIP) KWA support in WAMI Viewer
Modify vpFileDataSource to support reading a KWA index. This is the
first step in adding crude KWA support to WAMI Viewer. (Crude, because
WAMI Viewer is incredibly wedded to having individual image files and,
to some extent, to obtaining times and homographies separately from
image data. A proper implementation would replace all that mess with a
single "video" source that integrates everything, in order to provide a
clean solution for a source like KWA where all of the metadata is
tightly coupled with the image data.)

The second step will be to implement a vtkVgBaseImageSource that reads a
single frame from a KWA. (Unfortunately, this approach means that almost
none of the existing KWA code can be used, except as cut-and-paste into
the new image source.)