Add 'saved where' label to project editor
Add a label to vpProjectEditor that describes whether the project will
be saved on acceptance of the dialog, and if so, to what location. This
should help clarify that the project is not saved immediately on
selecting the location, as well as provide an additional reminder when
the project will not be saved.