Get rid of silly RegisterVpImageSource
Get rid of this "helper class" that exists to notionally simplify
registering image readers. In reality, wrapping a function call in a
struct ctor is confusing and not really beneficial, especially when
stripping out this layer of indirection makes the code not only easier
to follow, but actually shorter as well.