Implement VTK KWA reader
Implement a vtkVgBaseImageSource that can read a single frame from a
KWA. Add this to the set of readers registered with WAMI Viewer. In
conjunction with recent changes to vpFileDataSource, this allows WAMI
Viewer to use a KWA as its image source. (However, while time stamps
should be working, homographies are not currently supported. This might
be implemented at some point in the future.)