Implement vpKwiverImageLoader
Create a class that presents the interface (at least, the bit that we
care about) of a kwiver::vital::algo::image_io, but is implemented using
vpImageSourceFactory and vtkVgBaseImageSource. This gives us something
that can be used as a drop-in replacement for the image readers that
feed KWIVER (embedded pipelines, track improvement), but uses the same
(VTK-based) readers that WAMI Viewer uses internally.

This is particularly important for the recently added KWA support, which
relies on a kludged-up VTK reader to mangle "file names" into KWA frames
(and thus, obviously, doesn't work with KWIVER's readers).