Implement import/export of TOC filters
Rename existing actions to import and export "filters" to clarify what
filters are being imported or exported. Also implement import and export
of "classification" filters (for now, just track types).

The import currently is not very sophisticated; if the imported file
refers to a type name that is not currently known, the type is added,
without any regard to whether or not this is useful.

Conflicts:
	Applications/VpView/vpView.ui