Fix missing const in IdToPointXIter
Add missing const to dereference operator in IdToPointXIter. This struct
is attempting to be an iterator, but VS15 (a.k.a. VS 2017) gets in a
snit because there is no dereference operator that can be called on a
const instance of the struct. Fix this by making the operator const, as
it always should have been.