Refactor file reader utility classes
Rename the strangely named vpFile{Event,Track}IOImpl classes to
vpFile{Event,Track}Reader, and refactor them to have a pointer to the
underlying IO class, rather than static member functions that take the
IO class as the first parameter. This makes the code slightly cleaner
and easier to use.