Support traits, classifiers in VDF I/O
Modify vpVdfIO and vpVdfTrackIO to add support for track traits and
track classifiers files. This is still "broken" in that the classifiers
are reading fixed P/V/Os and not TOCs, but fixing that won't involve
changes to the VDF classes.

Support for region and attribute files still needs to be added.