Refactor find methods of vgTimeMap
Refactor how the find methods of vgTimeMap are implemented to avoid
needing to convert a const_iterator to an iterator, which is not
supported in Qt5. This requires that the implementation is templated on
not only the map's value type, but on the map pointer type (const or
not) and iterator type. In order to avoid users calling the helper with
crazy types, it is implemented as a private static method on the class
that is called (with appropriate types) from the public find methods.