Remove unneeded QCleanlooksStyle header
A long, long time ago (apparently, in the very first version ever
committed), WAMI Viewer forcibly set the widget style to Cleanlooks.
While it has not done so in a similarly long time, the include never
went away. Delete it now.

This will be one less stumbling block when we switch to Qt5.