Use UTF-16 literals
Replace escaped UTF-8 in string literals with escaped UTF-16, and use
UTF-16 literals (both are features new in C++11). Simplify conversion to
QString by introducing a function to emulate QStringLiteral (from Qt5),
and also use this in some opportune places in the files being modified.
Our version of QStringLiteral isn't quite as good, since it is just
constructing a QString from a UTF-16 literal using QString::fromUtf16,
but it is source compatible with the Qt5 version which will facilitate
porting later.