Teach vpFileTrackReader to read TOC's
Modify vpFileTrackReader::ReadTrackClassifiers to be able to read either
P/V/O's or TOC's. Refactor and unify some code to obtain track type
indices.

The reader assumes that a record of four values, of which the last three
are convertible to real numbers, is a P/V/O. Otherwise, it will attempt
to read the record as a TOC.