Start porting build system to Qt5
Modify CMake logic to find Qt5 instead of Qt4. Replace qt4 macros with
qt5 versions (mechanical change).

Note that THIS DOES NOT BUILD. Some source porting was done previously
to the extent that the code could still build with Qt4, but additional
source and build changes are required. It will likely be several commits
before things are building again.