Use Eigen instead of QMatrixNxN
Replace use of QMatrix3x3, QMatrix4x4 with Eigen matrices. This adds a
hard dependency on Eigen, however the API is nicer than VNL, and most
users are probably using KWIVER these days which already has such a
dependency. This is needed because Qt5 changes the underlying data type
of these classes from qreal (a.k.a. double) to float, which would
require us to change our conversion code (the current code would build
under Qt5 but cause a very nasty stack corruption!) and deal with a ton
of warnings, as well as costing precision. Also, Eigen conveniently lets
us force the data ordering which allows us to omit a bunch of
transpositions and/or element-by-element copying that were required with
the Qt types.

This also adds an interface include dependency on an external library
which might be used for headers only, which our old home-grown version
of usage requirements does not support. Address this by adding a new
vg_add_library which also creates an INTERFACE target for consuming only
the headers of a library, and use that with modern usage requirements to
propagate the dependency. This is a first step at hopefully eventually
ripping out much of our old home-grown usage requirements system and
replacing it with upstream usage requirements.