Use "native" usage requirements
Remove our home-grown system for specifying usage requirements (i.e. 'if
you use this library, also use these include paths') and replace with
use of CMake's native support. (Ours was first! ...but now that it is
upstream, we don't need ours any more.)

Fortuitously, this is mostly a matter of some small changes in our
install_headers function and replacing our wrapper functions with native
functions, so the changes are fairly minimal and non-invasive.

Unfortunately, this breaks the build (for now), because some headers
that need to be moc'd can't find required includes. This is because our
includes used to be directory-scope, but are now target-scope, and
qt5_wrap_cpp doesn't know about target-level includes, because the
target doesn't exist yet when it is used.

The solution will be to switch to automoc. This will be done in the next
commit.