Fix Unicode string literals on MSVC
Due to the combined quirkiness of how VS processes string literals and
how QStringLiteral (or more specifically, QT_UNICODE_LITERAL) is
implemented, Visual Studio clobbers characters that are outside the
local code page in certain uses of QStringLiteral. This happens because
QT_UNICODE_LITERAL does not actually force the literal to be UTF-16, but
rather prepends it with a UTF-16 literal, relying on concatenation
promotion to make the remaining string also UTF-16. However, by the time
that happens, Visual Studio has already corrupted the literal data.

The work-around is to prefix the actual literal (i.e. u"foo" instead of
"foo" inside of QStringLiteral).

Additionally, turn the warning about this (C4566) into an error, which
should hopefully prevent this from happening.