Extract and update qHash(vvTrackId)
Apparently, in Qt5 it is very complicated to forward declare qHash (at
least, MSVC didn't like my naïve attempt; it seems use of constexpr is
quirky on MSVC), so instead of trying, extract that function to a
separate header (now in vvIO) that just includes <QHashFunctions>. This
also accounts for changes to qHash between Qt4 and Qt5.

It seems this is not actually being used at the moment, but I don't
really want to get rid of it entirely.