Use full image dimensions when determining need for Y offset
An image shift is necessary so that the upper left is treated as the origin
if input images do not all have the same height. The logic doing the shift
was incorrectly doing a shift when the input images had the same height but
zooming in (or translating) had caused the reader (vtkVgGDALReader) to only
return a subset of the image.