Add generic event type
Create vvEvent; a simple type for conveying events along the lines of
vvTrack. This will be used as the base data type for creating a
frameworks event source.

(Note that, despite the pedigree of the vvData library, this new type
reflects a modern vision of events, and not the old fixed-vector event
descriptors from days of yore, a la vvDescriptor.)