Create vdfTrackOracleEventDataSource
Create a new frameworks data source that reads events using
track_oracle. This is added to the existing track_oracle frameworks data
source plugin.

In theory, this is fully functional; however, nothing is using it yet,
and (as of writing) track_oracle does not have any file formats that
actually implement the required schema.

For now, also disable the "generic" track_oracle formats, as our current
mechanism for determining if we want a track source or an event source
has no feasible way of making that distinction for these formats.