Move type lookup to base class
Move functions to look up type indices to vtkVgTypeRegistry from the
derived classes vtkVgTrackTypeRegistry and vtkVgEventTypeRegistry. This
makes the ability to look up an index from either the type ID or type
name available to users of all registry types. (Previously, lookup by
name was supported only for track types, and lookup by ID was supported
only for event types, which made for a bizarre inconsistency.)