Remove silly vidtk reference
Rename homography variables to not talk about "vidtk coordinates". While
this is not completely incorrect, vidtk is using "image" coordinates
(i.e. origin is top left, +Y is down), which are hardly unique to vidtk,
even in the context of our homographies (as our data can come from other
sources these days, which also tend to use image coordinates).

Accordingly, replace "vidtk" with "image", so that we are using the more
general (and arguably more accurate) terminology, and have one fewer
spurious reference to a completely internal toolkit which we are slowly
excising.