Relax required track schema
Modify the track_oracle format listing logic to accept readers that have
either a time value or a frame number (and not necessarily both), rather
than requiring a time value. Our own code already supported this, except
that, prior to this change, we wouldn't even consider formats that don't
at least potentially provide a time value. This should increase the
number of formats we potentially support.