Estimate track points when not provided
Add a helper function to vpTrackIO to estimate the location of a track
point from the bounding box, in case it is not available. Modify
vpVdfTrackIO to use this as needed. Modify vdfTrackOracleTrackDataSource
to set the track point to NaN when not available so that consumers (i.e.
vpVdfTrackIO) can tell that it is not available.

Note that "estimation", here, is just the "center" of the bounding box,
which is how we have derived these points as needed since time
immemorial. However, depending on the type of data, we sometimes use the
"true" center, and sometimes the middle of the bottom edge. The helper
function provides a QSettings-controlled mechanism for selecting which
method is used (defaulting to "true" center).