Don't crash if vdfDataReader::setSource fails
Modify vdfDataReader::failed to not crash if no source is set. Modify
vdfDataReader users (vpVdfTrackIO, vpVdfEventIO) to check for failure of
vdfSDataReader::setSource.

Historically, we expected failure to "never" happen, but it can happen
if we try to use a reader that does not provide a required interface
(e.g. an event source with a track reader).