Allow creating sources with requested interfaces
Add additional logic to vdfSourceService and vdfTrackOracleArchiveSource
to allow users to specify a list of desired interfaces for the source to
be created, and to use this list as a hint when deciding what source to
create. Modify users to take advantage of this functionality.

This works around a track_oracle limitation where the track and event
sources both equally believe they can read each other's files.