Fix issues building with Visual Studio
Refactor some loops to avoid a Visual Studio bug (that was fixed in
19.23, but Rusty really doesn't want to upgrade yet). In particular, the
somewhat odd-looking construct `decltype(+end)` is intended to strip
const-ness off the type (`end` is const) by performing an operation that
changing the expression from a reference to a prvalue (via the builtin
unary `+` operator), which should result in a non-const type. This works
in every compiler which implements C++11 (or even mostly does so; hello,
GCC 4.8!)... EXCEPT for Visual Studio prior to 19.23. The work-around is
slightly more verbose, but also less esoteric, and conforms to AAA
whereas the old code did not.

Also, add a missing include for std::max.