Make "private" headers possible
Add ability to have "exported" but not installed headers. This is
required for some libraries that need to use private headers of other
libraries, because ViViA is set up to only use a "fake install tree" to
resolve cross-library includes as a sanity check against forgetting to
install headers that are part of a library's ABI. Unfortunately, this
historically meant that we also could not use "private" headers across
libraries without installing them anyway (making them not exactly
"private").

Use this new ability to avoid installing "private" headers.

Also, switch to cmake_parse_arguments and get rid of the last vestiges
of our old, home-grown system.