Fix handling of "exported" per-config headers
Tweak vg_export_headers to create its build-tree wrappers in a way that
supports configuration-specific headers. Modify rules for vvVtkWidgets
so that "private" (non-installed) headers are generated in a
configuration-specific location when using a multi-configuration
generator.

This should correctly fix the issue that 514a988ba6cf attempted to fix,
but only succeeding in fixing for single-configuration generators.

(In practice, is is extremely unlikely that the headers in question
would have different content, but this is something of a work-around for
how AUTOUIC works.)