Revert "Kwiver recently discarded detected_object_type (#69)"
Revert commit 9a4b54076e3d588f2dcbec3ad0777e6e8d06ad51. The upstream
(KWIVER) changes caused issues due to conflation of detected object and
activity types, due to which class_map was refactored to be a base
class, and the original detected_object_type was reintroduced. Since
this means that consumers do not need to be modified, revert the prior
changes.