Tolerate some imprecision for KWA clips
Modify various bits of the KWA clip-finding logic to add some fuzz
factors. This should make the code tolerant of a small amount of
imprecision (which can happen, given that we are using floating point
numbers) when looking up matching clips and resolving padding. This
should fix these operations failing to produce correct results in some
cases. In particular, this could occur if a request is for a single
frame at either the start or end of the KWA.